backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
psycopg2-binary
langchain
langchain-core
langchain-groq
langchain-community
langgraph
tavily-python